requirements.txt
streamlit
pytesseract
pdf2image
docling
pymupdf4llm
mistralai
python-dotenv
